+++ modules/@apostrophecms/schema/ui/apos/logic/AposSchema.js
@@ -181,7 +181,7 @@ export default {
         this.populateDocData();
       }
     },
-    generation() {
+    generation(generated) {
       // repopulate the schema.
       this.populateDocData();
     },
@@ -331,6 +331,10 @@ export default {
     },
     highlight(fieldName) {
       return this.meta[fieldName]?.['@apostrophecms/schema:highlight'];
+    },
+    generateItemUniqueKey(field) {
+      return `${field.name}:${field._id ?? ''}:${this.modelValue?.data?._id ?? ''}`;
+
     }
   }
 };